Split txsender into files
